### 2018-09-05 #929 @whedon
[PRE REVIEW]: survxai: an R package for model agnostic explanations of survival models
**Submitting author:** @AleksandraDabrowska (<a href="http://orcid.org/0000-0002-5963-6260">Aleksandra Grudziąż</a>)
**Repository:** <a href="https://github.com/MI2DataLab/survxai" target ="_blank">https://github.com/MI2DataLab/survxai</a>
**Version:** 0.2.0
