### 2018-09-10 #929 @yochannah
@hiendn You are awesome! I'm still hunting for a second reviewer before I kick off the official review process, but in the meantime you can check out [the reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) and/or start looking at the software package.
