### 2018-09-20 MI2DataLab/survxai#18 @hiendn
Corrections to JOSS manuscript
Please make the corrections from the marked up file that has been attached. Please also conduct a thorough proof reading once you have made the corrections, since I may have missed some things.

[10.21105.joss.00961.pdf](https://github.com/MI2DataLab/survxai/files/2401202/10.21105.joss.00961.pdf)
