### 2018-09-23 MI2DataLab/survxai#18 @dirmeier
In addition to the corrections mentioned by @hiendn, please find a list of minor revisions and comments below:
- There are many typos and grammatical errors. Please make sure to remove them.
- Even though I understand your point, from the intro it does not get clear to someone not working in the field how survival analysis models and your tool are related to NNs or RFs, or rather it does not get enough emphasis. You motivate your problem with the interpretability of models that are "complex in structure", yet only show *local method* plots for a simple Cox proportional hazard model where intepretation of the coefficients is relatively simple, right? I understand that this is only a simple use case that could be generalized to more complex problems, but I believe you could maybe better motivate this, e.g. by showing a more elaborate use case, for instance when the Cox model is parametrized by an NN.  This is however only a comment and I might be wrong.
- Figure 1: why do you only describe the middle panel and not the others? E.g. what are the colors for `bili`?
- Hyperlink of Eleuteri paper does not work. Also check other hyperlinks please.
