### 2018-09-23 MI2DataLab/survxai#20 @dirmeier
Default methods
If implementing a `default` method for an S3 function, I believe you should use
```{r}
explain  <- function() UseMethod("explain")
```
instead of
```{r}
explain  <- explain.default
```
This would remove the second entry in the function's docu.
