### 2018-09-23 MI2DataLab/survxai#21 @dirmeier
Documentation
Please revise the following bullet points:
- Some of the functions have a `...` parameter which is documented as `other parameters`. Could you please explain what these are?
- In `plot.surv_ceteris_paribus_explainer` it does not get clear what the other colors than black represent.
- In `prediction_breakdown` what is function `surv_breakdown`?
