### 2018-09-25 MI2DataLab/survxai#22 @hiendn
Failure to install from the GitHub repo
Hi @AleksandraDabrowska,

Installation seems to be fine with `install.packages('survxai')`.

However, I also tried to install the package using: `devtools::install_github("MI2DataLab/survxai")`, which did not work!

I am currently running R on a MacBook pro, version `R version 3.3.2 (2016-10-31) -- "Sincere Pumpkin Patch"`.

The full installation output is given below:

```
Downloading GitHub repo MI2DataLab/survxai@master
from URL https://api.github.com/repos/MI2DataLab/survxai/zipball/master
Installing survxai
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/pec_2.5.4.tgz'
Content type 'application/x-gzip' length 298313 bytes (291 KB)
==================================================
downloaded 291 KB

Installing pec
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/rms_5.1-2.tgz'
Content type 'application/x-gzip' length 1177171 bytes (1.1 MB)
==================================================
downloaded 1.1 MB

Installing rms
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/Hmisc_4.1-1.tgz'
Content type 'application/x-gzip' length 1795943 bytes (1.7 MB)
==================================================
downloaded 1.7 MB

Installing Hmisc
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/acepack_1.4.1.tgz'
Content type 'application/x-gzip' length 73287 bytes (71 KB)
==================================================
downloaded 71 KB

Installing acepack
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba47c9de3c/acepack'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘acepack’ ...
* DONE (acepack)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/Formula_1.2-2.tgz'
Content type 'application/x-gzip' length 160779 bytes (157 KB)
==================================================
downloaded 157 KB

Installing Formula
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8abae901b16/Formula'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘Formula’ ...
* DONE (Formula)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/htmlTable_1.11.1.tgz'
Content type 'application/x-gzip' length 238947 bytes (233 KB)
==================================================
downloaded 233 KB

Installing htmlTable
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba5214b7a7/htmlTable'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘htmlTable’ ...
* DONE (htmlTable)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/latticeExtra_0.6-28.tgz'
Content type 'application/x-gzip' length 2071651 bytes (2.0 MB)
==================================================
downloaded 2.0 MB

Installing latticeExtra
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba255a32c2/latticeExtra'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘latticeExtra’ ...
* DONE (latticeExtra)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba40da5995/Hmisc'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘Hmisc’ ...
* DONE (Hmisc)
Skipping install of 'htmlTable' from a cran remote, the SHA1 (1.11.1) has not changed since last install.
  Use `force = TRUE` to force installation
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/multcomp_1.4-8.tgz'
Content type 'application/x-gzip' length 626582 bytes (611 KB)
==================================================
downloaded 611 KB

Installing multcomp
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/TH.data_1.0-8.tgz'
Content type 'application/x-gzip' length 5200584 bytes (5.0 MB)
==================================================
downloaded 5.0 MB

Installing TH.data
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba6e6a96bc/TH.data'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘TH.data’ ...
* DONE (TH.data)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba68769de8/multcomp'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘multcomp’ ...
* DONE (multcomp)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/polspline_1.1.12.tgz'
Content type 'application/x-gzip' length 635021 bytes (620 KB)
==================================================
downloaded 620 KB

Installing polspline
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba497277e7/polspline'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘polspline’ ...
* DONE (polspline)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba4bb4d4ca/rms'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘rms’ ...
* DONE (rms)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/timereg_1.9.1.tgz'
Content type 'application/x-gzip' length 803369 bytes (784 KB)
==================================================
downloaded 784 KB

Installing timereg
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba38371592/timereg'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘timereg’ ...
* DONE (timereg)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba3b8446b6/pec'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘pec’ ...
* DONE (pec)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/survminer_0.4.1.tgz'
Content type 'application/x-gzip' length 2659127 bytes (2.5 MB)
==================================================
downloaded 2.5 MB

Installing survminer
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/cmprsk_2.2-7.tgz'
Content type 'application/x-gzip' length 70301 bytes (68 KB)
==================================================
downloaded 68 KB

Installing cmprsk
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba59223d72/cmprsk'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘cmprsk’ ...
* DONE (cmprsk)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/ggpubr_0.1.6.tgz'
Content type 'application/x-gzip' length 1499409 bytes (1.4 MB)
==================================================
downloaded 1.4 MB

Installing ggpubr
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/cowplot_0.9.2.tgz'
Content type 'application/x-gzip' length 2441681 bytes (2.3 MB)
==================================================
downloaded 2.3 MB

Installing cowplot
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba47435875/cowplot'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘cowplot’ ...
* DONE (cowplot)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/ggrepel_0.7.0.tgz'
Content type 'application/x-gzip' length 675838 bytes (659 KB)
==================================================
downloaded 659 KB

Installing ggrepel
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba79005c08/ggrepel'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘ggrepel’ ...
* DONE (ggrepel)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/ggsci_2.8.tgz'
Content type 'application/x-gzip' length 2959966 bytes (2.8 MB)
==================================================
downloaded 2.8 MB

Installing ggsci
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba744949b7/ggsci'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘ggsci’ ...
* DONE (ggsci)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/ggsignif_0.4.0.tgz'
Content type 'application/x-gzip' length 40583 bytes (39 KB)
==================================================
downloaded 39 KB

Installing ggsignif
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba79554537/ggsignif'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘ggsignif’ ...
* DONE (ggsignif)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba7f264308/ggpubr'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘ggpubr’ ...
* DONE (ggpubr)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/maxstat_0.7-25.tgz'
Content type 'application/x-gzip' length 152607 bytes (149 KB)
==================================================
downloaded 149 KB

Installing maxstat
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/exactRankTests_0.8-29.tgz'
Content type 'application/x-gzip' length 107610 bytes (105 KB)
==================================================
downloaded 105 KB

Installing exactRankTests
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba256a0f7e/exactRankTests'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘exactRankTests’ ...
* DONE (exactRankTests)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba19c7af4a/maxstat'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘maxstat’ ...
* DONE (maxstat)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/survMisc_0.5.4.tgz'
Content type 'application/x-gzip' length 777101 bytes (758 KB)
==================================================
downloaded 758 KB

Installing survMisc
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/km.ci_0.5-2.tgz'
Content type 'application/x-gzip' length 47684 bytes (46 KB)
==================================================
downloaded 46 KB

Installing km.ci
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba39329abf/km.ci'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘km.ci’ ...
* DONE (km.ci)
trying URL 'https://cran.rstudio.com/bin/macosx/mavericks/contrib/3.3/KMsurv_0.1-5.tgz'
Content type 'application/x-gzip' length 123065 bytes (120 KB)
==================================================
downloaded 120 KB

Installing KMsurv
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba659e878c/KMsurv'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘KMsurv’ ...
* DONE (KMsurv)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba68a70594/survMisc'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘survMisc’ ...
* DONE (survMisc)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba7f075b99/survminer'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

* installing *binary* package ‘survminer’ ...
* DONE (survminer)
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD INSTALL  \
  '/private/var/folders/_m/1phfpjh17qd22dfnmcs0pprn_h6t1t/T/RtmpcPzxDd/devtools8aba19e35532/MI2DataLab-survxai-0867032'  \
  --library='/Users/hnguyen/Library/R/3.3/library' --install-tests

ERROR: dependency ‘breakDown’ is not available for package ‘survxai’
* removing ‘/Users/hnguyen/Library/R/3.3/library/survxai’
Installation failed: Command failed (1)
```
