### 2018-09-25 MI2DataLab/survxai#22 @hiendn
Could you please put a comment in the paper regarding installation instructions from CRAN. That is, state that you can install using the command `install.packages('survxai')`.
