### 2018-10-08 #961 @yochannah
Quick status check - looking at the original repo, issues [18](https://github.com/MI2DataLab/survxai/issues/18) to [21[(https://github.com/MI2DataLab/survxai/issues/21) need to be addressed before we can take this one forward, right?
