### 2018-10-08 #961 @dirmeier
Yeah, exactly.  From my side these two and [20](https://github.com/MI2DataLab/survxai/issues/20). All of them are minor changes.
