### 2018-10-08 #961 @hiendn
@yochannah that's correct, resolving [22](https://github.com/MI2DataLab/survxai/issues/22) would be good as well, if possible.
