### 2018-10-22 877aabb
Add UseMethod to explain function #20
