+++ man/explain.Rd
@@ -1,16 +1,16 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/explain.R
-\name{explain.default}
-\alias{explain.default}
+\name{explain}
 \alias{explain}
+\alias{explain.default}
 \title{Create Survival Model Explainer}
 \usage{
-explain.default(model, data = NULL, y, times = NULL,
-  predict_function = yhat, link = I, label = tail(class(model), 1),
-  ...)
-
 explain(model, data = NULL, y, times = NULL, predict_function = yhat,
   link = I, label = tail(class(model), 1), ...)
+
+\method{explain}{default}(model, data = NULL, y, times = NULL,
+  predict_function = yhat, link = I, label = tail(class(model), 1),
+  ...)
 }
 \arguments{
 \item{model}{object - a survival model to be explained}
