+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(explain,default)
 S3method(plot,surv_ceteris_paribus_explainer)
 S3method(plot,surv_explainer)
 S3method(plot,surv_model_performance_explainer)
@@ -13,7 +14,6 @@ S3method(print,surv_prediction_breakdown_explainer)
 S3method(print,surv_variable_response_explainer)
 export(ceteris_paribus)
 export(explain)
-export(explain.default)
 export(model_performance)
 export(prediction_breakdown)
 export(theme_mi2)
