### 2018-10-22 MI2DataLab/survxai#20 @agosiewska
@dirmeier Thanks, I changed it according to your suggestion.
