+++ R/variable_response.R
@@ -1,13 +1,12 @@
 #' @title Variable response for survival models
 #'
-#' @description Function \code{variable response} calculates the expected output condition on a selected variable.
+#' @description Function \code{variable_response} calculates the expected output condition on a selected variable.
 #'
 #' @param explainer an object of the class 'surv_explainer'.
 #' @param variable a character with variable name.
 #' @param type a character - type of the response to be calculated.
 #' Currently following options are implemented: 'pdp' for Partial Dependency.
 #' @param link a function - a link function that shall be applied to raw model predictions. This will be inherited from the explainer.
-#' @param ... other parameters
 #'
 #' @examples
 #' \donttest{
@@ -26,7 +25,7 @@
 #' }
 #' @export
 
-variable_response <- function(explainer, variable, type = "pdp", link = explainer$link, ...){
+variable_response <- function(explainer, variable, type = "pdp", link = explainer$link){
   if (!("surv_explainer" %in% class(explainer))) stop("The variable_response() function requires an object created with explain() function from survxai package.")
   if (is.null(explainer$data)) stop("The variable_response() function requires explainers created with specified 'data' parameter.")
 
