+++ R/print_variable_response.R
@@ -1,7 +1,7 @@
 #' Variable Response Print
 #'
 #' @param x the model of 'surv_variable_response_explainer' class
-#' @param ... other parameters
+#' @param ... further arguments passed to or from other methods
 #' 
 #' @return a data frame
 #'
