+++ R/plot_prediction_breakdown.R
@@ -3,7 +3,7 @@
 #' @description Function plot for surv_breakdown object visualise estimated survival curve of mean probabilities in chosen time points.
 #'
 #' @param x an object of class "surv_prediction_breakdown_explainer"
-#' @param ... other arguments, for example additional object of class "surv_prediction_breakdown_explainer"
+#' @param ... optional, additional object of class "surv_prediction_breakdown_explainer"
 #' @param numerate logical; indicating whether we want to number curves
 #' @param lines logical; indicating whether we want to add lines on chosen time point or probability
 #' @param lines_type a type of line; see http://sape.inf.usi.ch/quick-reference/ggplot2/linetype
