+++ tests/testthat/test_plot_model_performance.R
@@ -4,6 +4,6 @@ source("objects_for_tests.R")
 
 test_that("Output", {
   expect_is(plot(mp_cph), "gg")
-  expect_is(plot(mp_cph, model_performance(surve_cph2, data = pbcTest)), "gg")
+  expect_is(plot(mp_cph, model_performance(surve_cph2)), "gg")
   expect_is(plot(mp_cph, mp_cph_artificial), "gg")
 })
