### 2018-10-24 7c7cb50
Update documentation #21
