+++ R/prediction_breakdown.R
@@ -7,7 +7,7 @@
 #' @param observation a new observation to explain
 #' @param time a time point at which variable contributions are computed. If NULL median time is taken.
 #' @param prob a survival probability at which variable contributions are computed
-#' @param ... other parameters corresponding to arguments from \code{broken} function from \code{breakDown} package. See https://github.com/pbiecek/breakDown/blob/master/R/break_agnostic.R for details
+#' @param ... other parameters corresponding to arguments from \code{\link[breakDown]{broken}} function from \code{breakDown} package. See https://github.com/pbiecek/breakDown/blob/master/R/break_agnostic.R for details
 #'
 #' @return An object of class surv_prediction_breakdown_explainer
 #'
@@ -25,7 +25,7 @@
 #'                   return(prob)
 #'                   }
 #' cph_model <- cph(Surv(years, status)~., data=pbcTrain, surv=TRUE, x = TRUE, y=TRUE)
-#' surve_cph <- explain(model = cph_model, data = pbcTest[,-c(1,5)], 
+#' surve_cph <- explain(model = cph_model, data = pbcTest[,-c(1,5)],
 #'                     y = Surv(pbcTest$years, pbcTest$status), predict_function = predict_times)
 #' broken_prediction <- prediction_breakdown(surve_cph, pbcTest[1,-c(1,5)])
 #' }
@@ -45,12 +45,12 @@ prediction_breakdown <- function(explainer, observation, time = NULL, prob = NUL
   res<- broken(model = explainer$model,
                   new_observation = observation,
                   data = explainer$data,
-                  predict.function = new_pred, 
+                  predict.function = new_pred,
                   ...)
   options(warn = oldw)
 
   class(res) <- "data.frame"
-  
+
   intercept <- res$contribution[res$variable_name=="Intercept"]
   observ <- res$contribution[res$variable=="final_prognosis"]
 
@@ -92,34 +92,34 @@ prediction_breakdown <- function(explainer, observation, time = NULL, prob = NUL
 predict_fun <- function(prob, time, explainer){
   if (is.null(prob)) {
     if (is.null(time)) time <- median(explainer$times)
-    
+
     new_pred <- function(model, data){
       explainer$predict_function(model, data, times = time)
     }
   } else {
     times_sorted <- sort(explainer$times)
-    
+
     find_time <- function(x){
       tim <- (x < prob)
       index <- c(min(which(tim == TRUE)) -1, min(which(tim == TRUE)))
       closest_times <- times_sorted[index]
       weighted.mean(closest_times, x[index])
     }
-    
+
     new_pred <- function(model, data){
       probabilities <- explainer$predict_function(model, data, times = explainer$times)
       probabilities <- as.data.frame(probabilities)
-      
+
       res <- apply(probabilities, MARGIN = 1, FUN = find_time)
       res <- na.omit(res)
       return(res)
-      
+
     }
-    
+
     npred <- new_pred(explainer$model, explainer$data)
     message("Number of observations with prob > ", prob, ": ", nrow(explainer$data) - length(npred))
   }
-  
+
   return(new_pred)
 }
 
