+++ R/plot_model_performance.R
@@ -3,7 +3,7 @@
 #' @description Function plot for surv_model_performance object.
 #'
 #' @param x object of class "surv_model_performance"
-#' @param ... optional, additional object of class "surv_model_performance_explainer"
+#' @param ... optional, additional objects of class "surv_model_performance_explainer"
 #'
 #' @import ggplot2
 #'
@@ -18,7 +18,7 @@
 #'                   return(prob)
 #'                   }
 #' cph_model <- cph(Surv(years, status)~., data=pbcTrain, surv=TRUE, x = TRUE, y=TRUE)
-#'surve_cph <- explain(model = cph_model, data = pbcTest[,-c(1,5)], 
+#'surve_cph <- explain(model = cph_model, data = pbcTest[,-c(1,5)],
 #'                     y = Surv(pbcTest$years, pbcTest$status), predict_function = predict_times)
 #' mp_cph <- model_performance(surve_cph, data = pbcTest)
 #' plot(mp_cph)
