+++ misc/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'survxai: an R package for model agnostic explanations of survival models'
+title: 'survxai: an R package for model-agnostic explanations of survival models'
 authors:
 - affiliation: 1, 2
   name: Aleksandra Grudziaz
@@ -10,7 +10,7 @@ authors:
 - affiliation: 1, 2
   name: Przemyslaw Biecek
   orcid: 0000-0001-8423-1823
-date: "3 August 2018"
+date: "23 August 2018"
 output: pdf_document
 bibliography: paper.bib
 tags:
@@ -20,63 +20,63 @@ tags:
 - interpretable machine learning
 affiliations:
 - index: 1
-  name: Faculty of Mathematics, Informatics, and Mechanics, University of Warsaw
-- index: 2
   name: Faculty of Mathematics and Information Science, Warsaw University of Technology
+- index: 2
+  name: Faculty of Mathematics, Informatics, and Mechanics, University of Warsaw
 ---
 
 # Introduction
 
 <!-- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience-->
 
-Predictive models are widely used in the supervised machine learning. Three most common classes of such models are: *regression models*, here the target variable is continuous numeric, *classification models* here the target variable is binary or categorical and *survival models* here the target is some censored variable. Common examples of censored variables are time to death (but some cases lived for x years and are still alive), cessation of service by the customer or failure of a machine component.
+Predictive models are widely used in supervised machine learning. Three most common classes of such models are: *regression models*, where the target variable is continuous numeric, *classification models*, where the target variable is binary or categorical and *survival models*, where the target is some censored variable. Common examples of censored variables are times to death (but some cases lived for x years and are still alive), cessation of service by customer, or failure of machine components.
 
-Modern predictive models are often complex in structure. Think about neural networks [@ELEUTERI2003855] or random forest [@randomForestSRC]. Such models may be described by thousands of coefficients. Often such flexibility leads to high performance, but makes these models opaque, hard to understand. It is acceptable in cases in which only the model accuracy is important, but in cases that involve human decisions it may not be enough. To trust model predictions one needs to see which features are important and how model predictions would change if some feature was changed. 
+Modern predictive models are often complex in structure; for example neural networks [@ELEUTERI2003855] or random forest [@randomForestSRC]. Such sophisticated tools are also more and more often used in survival analysis. These models may be described by thousands of coefficients. Often such flexibility leads to high performance, but makes models opaque and hard to understand. This is acceptable in cases where only the model accuracy is important, but in cases that involve human decisions, it may not be informative enough. To trust model predictions one needs to see which features are important and how model predictions would change if some feature was changed.
 
-The area of model interpretability or explanability gains quickly attention of machine learning experts. Understanding of complex models leads not only to higher trust in model predictions but also to better models. Better means that they are more robust and maintain high accuracy on validation data. See examples in `DALEX`  [@2018arXiv180608915B] or `iml` [@Molnar2018] R packages. 
+The area of model interpretability or explanability has quickly gained the attention of machine learning experts. Understanding of complex models leads not only to higher trust in model predictions but also better models. Better, means that the models are more robust and obtains higher accuracy on validation data. See examples in the `DALEX`  [@2018arXiv180608915B] or `iml` [@Molnar2018] R packages.
 
-Existing tools for model agnostic explanations are focused on regression models and classification problems as in both cases model predictions may be summarised by a single number. Survival models require different approach as predictions are in a form of survival curves. Demand for such explainers leads to some model specific solutions, like `iSurvive` introduced by [@pmlr-v70-dempsey17a] for continuous time hidden Markov model. Yet, we are lacking of model agnostic tools for survival models.
[…]
