+++ misc/paper.bib
@@ -3,21 +3,19 @@
     author = {Biecek, Przemysław},
     year = {2018},
     note = {R package version 0.2.1},
-    url = {https://CRAN.R-project.org/package=ceterisParibus}
+    url = {http://CRAN.R-project.org/package=ceterisParibus}
   }
 
 @article{2018arXiv180608915B,
-   author = {{Biecek}, Przemysław},
+	author = {{Biecek}, Przemysław},
     title = "{DALEX: explainers for complex predictive models}",
-  journal = {ArXiv e-prints},
-archivePrefix = "arXiv",
-   eprint = {1806.08915},
- primaryClass = "stat.ML",
- keywords = {Statistics - Machine Learning, Computer Science - Artificial Intelligence, Computer Science - Machine Learning, Statistics - Applications},
-     year = 2018,
-    month = jun,
-   adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180608915B},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    url = {http://arxiv.org/abs/1806.08915},
+	eprint = {1806.08915},
+	primaryClass = "stat.ML",
+	keywords = {Statistics - Machine Learning, Computer Science - Artificial Intelligence, Computer Science - Machine Learning, Statistics - Applications},
+    year = 2018,
+	adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180608915B},
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 
@@ -26,7 +24,7 @@ archivePrefix = "arXiv",
   author={Collett, David},
   isbn={9781498731690},
   series={Chapman \& Hall/CRC Texts in Statistical Science},
-  url={https://books.google.pl/books?id=Okf7CAAAQBAJ},
+  url={http://books.google.pl/books?id=Okf7CAAAQBAJ},
   year={2015},
   publisher={CRC Press}
 }
@@ -43,11 +41,12 @@ archivePrefix = "arXiv",
 @Article{randomForestSRC,
   title = {Random survival forests},
   author = {Ishwaran, Hemant and Kogalur, Udaya B. and Blackstone, Eugene H. and Lauer, Michael S.},
-  journal = {Ann. Appl. Statist.},
+  journal = {The Annals of Applied Statistics},
   year = {2008},
   volume = {2},
   number = {3},
   pages = {841--860},
+  doi = {doi:10.1214/08-AOAS169},
   url = {http://arXiv.org/abs/0811.1645v1},
   pdf = {http://arxiv.org/pdf/0811.1645},
 }
@@ -61,8 +60,8 @@ archivePrefix = "arXiv",
   year = {2003}, 
   note = {Advances in Neural Networks Research: IJCNN '03},
   issn = {0893-6080},
-  doi = {https://doi.org/10.1016/S0893-6080(03)00098-4},
-  url = {http://www.sciencedirect.com/science/article/pii/S0893608003000984},
+  doi = {10.1016/S0893-6080(03)00098-4},
+  url = {http://doi.org/10.1016/S0893-6080(03)00098-4},
   author = {Eleuteri, Antonio and Tagliaferri, Roberto and Milano, Leopoldo and De Placido, Sabino and De Laurentiis, Michele},
   keywords = {Survival analysis, Conditioning probability estimation, Neural networks, Bayesian learning, MCMC methods}
 }
@@ -72,7 +71,7 @@ archivePrefix = "arXiv",
   title = {{pdp: An R Package for Constructing Partial Dependence Plots}},
   year = {2017},
   journal = {{The R Journal}},
[…]
