### 2018-10-30 355851a
Add Przemek's corrections
