+++ misc/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'survxai: an R package for model-agnostic explanations of survival models'
+title: 'survxai: an R package for structure-agnostic explanations of survival models'
 authors:
 - affiliation: 1, 2
   name: Aleksandra Grudziaz
@@ -31,13 +31,13 @@ affiliations:
 
 Predictive models are widely used in supervised machine learning. Three most common classes of such models are: *regression models*, where the target variable is continuous numeric, *classification models*, where the target variable is binary or categorical and *survival models*, where the target is some censored variable. Common examples of censored variables are times to death (but some cases lived for x years and are still alive), cessation of service by customer, or failure of machine components.
 
-Modern predictive models are often complex in structure; for example neural networks [@ELEUTERI2003855] or random forest [@randomForestSRC]. Such sophisticated tools are also more and more often used in survival analysis. These models may be described by thousands of coefficients. Often such flexibility leads to high performance, but makes models opaque and hard to understand. This is acceptable in cases where only the model accuracy is important, but in cases that involve human decisions, it may not be informative enough. To trust model predictions one needs to see which features are important and how model predictions would change if some feature was changed.
+Modern survival models are often complex in structure; for example survival neural networks [@ELEUTERI2003855] or survival random forest [@randomForestSRC]. These models may be described by thousands of coefficients. Often such flexibility leads to high performance, but makes models opaque and hard to understand. This is acceptable in cases where only the model accuracy is important, but in cases that involve human decisions, it may not be informative enough. To trust model predictions one needs to see which features are important and how model predictions would change if some feature was changed.
 
 The area of model interpretability or explanability has quickly gained the attention of machine learning experts. Understanding of complex models leads not only to higher trust in model predictions but also better models. Better, means that the models are more robust and obtains higher accuracy on validation data. See examples in the `DALEX`  [@2018arXiv180608915B] or `iml` [@Molnar2018] R packages.
 
-Existing tools for model agnostic explanations are focused on regression models and classification problems, as in both cases model predictions that may be summarised by a single number. Survival models require different approaches as predictions are in a form of survival curves. Demand for such explainers has led to some model specific solutions, like `iSurvive` [@pmlr-v70-dempsey17a] for continuous time hidden Markov models. Yet, there is currently a lack of model agnostic tools for survival models.
+Existing tools for model agnostic explanations are focused on regression models and classification problems, as in both cases model predictions that may be summarised by a single number. Survival models require different approach as predictions are in a form of survival functions. Demand for such explainers has led to some model specific solutions, like `iSurvive` [@pmlr-v70-dempsey17a] for continuous time hidden Markov models. Yet, there is currently a lack of structure agnostic tools for survival models.
 
-The `survxai` fills this gap. This R package is designed to deliver local and global explanations for survival models, in a model-agnostic fashion. In the package documentation we demonstrate examples for survival random forest models and for Cox models.
+The `survxai` fills this gap. This R package is designed to deliver local and global explanations for survival models, in a structure-agnostic fashion. In the package documentation we demonstrate examples for survival random forest models and for Cox models.
 The `survxai` package consists of new implementations and visualisations of explainers, designed for survival models. Functions are well documented and the package is supplemented with unit tests, and illustrations.
 Regardless of the complexity of the model, the methods implemented in the `survxai` package maintain a certain level of interpretability, important in medical applications [@collett2015modelling], churn analysis [@lu2003modeling] and others.
 
@@ -45,30 +45,30 @@ Regardless of the complexity of the model, the methods implemented in the `survx
 <!-- A clear statement of need that illustrates the purpose of the software-->
 # Explanations of survival models
  
[…]
