### 2018-10-30 MI2DataLab/survxai#18 @agosiewska
Thank you for your comments.
We've made the corrections and @AleksandraDabrowska rebuilded the pdf.
https://github.com/openjournals/joss-reviews/issues/961
