### 2018-10-30 MI2DataLab/survxai#21 @agosiewska
Thank you for your feedback.
We have updated the documentation.
