### 2018-10-30 MI2DataLab/survxai#22 @agosiewska
The comment regarding installation is now in the paper :)
https://github.com/openjournals/joss-reviews/issues/961
