### 2018-11-03 #961 @AleksandraDabrowska
@hiendn @dirmeier  Thank you very much for reviewing and for your kind words! :)
Should we now issue a DOI for our archive, as it is written in "The review process" section (https://joss.readthedocs.io/en/latest/submitting.html?fbclid=IwAR04SV6oxRM9B3mPL8NXhMhzToU_c3PwwcuiATwttKWaty2rEm8X8aZ19TM)?
