### 2018-11-04 #961 @yochannah
@AleksandraDabrowska Looking good! Before generating the archive, though, there is one issue with the manuscript I'd like to fix. Figures 3 and 4 are interspersed with the references in the most recent proof, but I'm reasonably sure they should be up in the main paper text. Can you fix this? Once this is fixed, please do generate an archive and share the URL in this thread!! 🎉
