### 2018-11-05 7e3fa4d
Update misc/paper.md
