### 2018-11-05 #961 @yochannah
The figures look better! Thanks @AleksandraDabrowska
