### 2018-11-06 #961 @arfon
@hiendn, @dirmeier - many thanks for your reviews here, and to @yochannah for editing this submission ✨

@AleksandraDabrowska - your paper is now accepted into JOSS :zap: :rocket: :boom:
